feat: Phase 4b — finish the SDK cutover (option-a finish)
This is the final cleanup of the SDK cutover port: prune the desktop and
agent-core dead code that fell out of scope once `apps/daemon` took over
task execution and the SDK adapter replaced the PTY adapter.

Desktop side:
- Delete `electron-options.ts` (dead `createElectronTaskManagerOptions`
  + PTY-era TaskManager wiring — desktop no longer creates a TaskManager;
  the daemon does).
- Delete `config-generator.ts` (was the desktop wrapper around
  agent-core's `resolveTaskConfig` "one brain"; daemon now generates
  per-task config inside `task-config-builder.onBeforeStart`).
- Delete `environment-builder.ts` (only `cleanupVertexServiceAccountKey`
  was still used; extracted to `opencode/vertex-cleanup.ts`).
- Extract `stopDevBrowserServer` to `opencode/dev-browser-shutdown.ts`
  (the only other still-used export from the deleted `electron-options.ts`).
- Drop the pre-emptive `generateOpenCodeConfig()` call from Slack MCP
  OAuth — the daemon writes the config lazily on task start now, so
  pre-staging it from the desktop main process is unnecessary for the
  OAuth handshake itself.
- Slim `opencode/index.ts` barrel to only the surface that actually has
  consumers.
- Delete the orphan `config-generator.integration.test.ts` (it tested the
  deleted desktop wrapper).

Agent-core side:
- Delete `opencode/resolve-task-config.ts` and its test (only consumer
  was the deleted desktop config-generator).
- Delete `opencode/cli-args.ts` (`buildCliArgs`) — the SDK adapter uses
  `session.prompt`, not CLI args. Z.AI provider normalisation for SDK
  calls already lives in `model-runtime-mapping.normalizeSelectedModelForSdk`.
- Drop public re-exports of `resolveTaskConfig`, `ResolveTaskConfigOptions`,
  `ResolvedTaskConfig`, `buildCliArgs`, `BuildCliArgsOptions` from
  `agent-core/src/index.ts`.
- Collapse `getOpenCodeAuthPath` to delegate to the XDG-aware
  `getOpenCodeAuthJsonPath` (per plan decision #6 — stops two helpers
  computing different paths when `XDG_DATA_HOME` is set).
- Update `apps/daemon/src/task-config-builder.ts` to use
  `getOpenCodeAuthJsonPath` instead of the legacy non-XDG helper.
- Strip the dead PTY-era fields (`getCliCommand`, `buildEnvironment`,
  `buildCliArgs`) from `AdapterOptions` and `TaskAdapterOptions`. The
  SDK adapter never read them — they were back-compat placeholders kept
  during the Phase 1b/2 transition and now removed. `TaskManager` no
  longer wraps `buildCliArgs` per task; `TaskManagerOptions.adapterOptions`
  is now the simple `AdapterOptions`.
- Remove the `getCliCommand` / `buildEnvironment` / `buildCliArgs`
  helpers from `apps/daemon/src/task-config-builder.ts` and the matching
  daemon `task-service.ts` wiring — they were the last callers of the
  back-compat fields above.

Test/mocks updated:
- `packages/agent-core/tests/unit/opencode/task-manager.test.ts` — drop
  the now-removed `buildCliArgs` / `buildEnvironment` / `getCliCommand`
  fields from the adapter-options sample.
- `packages/agent-core/tests/unit/opencode/config-generator.test.ts` —
  drop the `buildCliArgs` describe block; comment redirects to
  `model-runtime-mapping.unit.test.ts` for SDK-side normalisation tests.
- `apps/daemon/__tests__/unit/task-service-parity.test.ts` — refresh the
  agent-core mock to drop `resolveTaskConfig`, `buildOpenCodeEnvironment`,
  `resolveCliPath`, `getOpenCodeAuthPath`, and add `buildProviderConfigs`
  / `getOpenCodeAuthJsonPath` so the new daemon wiring resolves.

Gate (Node 24.14.1):
- `pnpm typecheck` — pass (all four workspaces)
- `pnpm lint:eslint` — pass
- `pnpm format:check` — pass
- `pnpm -F @accomplish_ai/agent-core test` — 582 passed / 1 skipped
- `pnpm -F @accomplish/desktop test:unit` — 281 passed
- `pnpm -F @accomplish/web test:unit` — 304 passed
- `pnpm -F @accomplish/daemon test` — 88 passed / 3 failed; the 3 failing
  cases (`task-service-parity.test.ts` attachments + listTasks) fail on
  `main` too under Node 24 and are unrelated to this port.

Realistic-scope note vs. the plan: the plan listed deletion of the full
"15+ CLI-era files" set under `packages/agent-core/src/opencode/`
(accomplish-runtime, auth-slack-mcp, cli-path-utils, cli-resolver,
config-generator-types, config-provider-context, 7× config-providers-*,
generator-mcp, message-{attachments,batcher,sanitization}, system-prompt-*).
Those files are still load-bearing for the OSS daemon's per-task
`opencode.json` generation and for `message-processor.ts`. Commercial
1a320029 was able to drop them because its adapter delegates entirely
to the SDK provider registry; OSS still owns config generation in the
daemon, so they remain. A follow-up "agent-core opencode/ convergence"
PR can do the deeper restructure once the daemon's config path moves
to the SDK provider registry too. The aspirational goal in the plan
was "perfect"; what landed in this port is "complete and not leaving
dead PTY-era code on the desktop or in the public surface."
