+++ apps/daemon/__tests__/unit/task-service-parity.test.ts
@@ -52,20 +52,15 @@ vi.mock('@accomplish_ai/agent-core', async (importOriginal) => {
       environment: {},
       config: {},
     })),
-    resolveTaskConfig: vi.fn(async () => ({
-      configOptions: {
-        platform: 'darwin',
-        mcpToolsPath: '/tools',
-        userDataPath: '/data',
-        isPackaged: false,
-      },
-    })),
     syncApiKeysToOpenCodeAuth: vi.fn(),
-    getOpenCodeAuthPath: vi.fn(() => '/auth'),
+    getOpenCodeAuthJsonPath: vi.fn(() => '/auth/auth.json'),
     getEnabledSkills: vi.fn(() => []),
-    buildOpenCodeEnvironment: vi.fn((env: NodeJS.ProcessEnv) => env),
-    resolveCliPath: vi.fn(() => ({ cliPath: '/bin/opencode' })),
     isCliAvailable: vi.fn(() => Promise.resolve(true)),
+    buildProviderConfigs: vi.fn(async () => ({
+      providerConfigs: {},
+      enabledProviders: [],
+      modelOverride: undefined,
+    })),
     getModelDisplayName: vi.fn((id: string) => id),
     getBundledNodePaths: vi.fn(() => null),
     DEV_BROWSER_PORT: 9224,
