+++ apps/desktop/src/main/opencode/slack-auth/index.ts
@@ -1,6 +1,5 @@
 import { randomUUID } from 'node:crypto';
 import { shell } from 'electron';
-import { generateOpenCodeConfig } from '../config-generator';
 import {
   discoverOAuthMetadata,
   discoverOAuthProtectedResourceMetadata,
@@ -36,8 +35,11 @@ export class SlackMcpOAuthFlow {
     let callbackServer: OAuthCallbackServer | null = null;
 
     try {
-      await generateOpenCodeConfig();
-
+      // Phase 4b of the OpenCode SDK cutover port removed the pre-emptive
+      // `generateOpenCodeConfig()` call here. Under the SDK architecture the
+      // daemon owns task-runtime config generation and writes it lazily when
+      // a task starts; pre-staging an `opencode.json` from the desktop main
+      // process is no longer required for the Slack OAuth handshake itself.
       const metadata = await discoverOAuthMetadata(OPENCODE_SLACK_MCP_SERVER_URL);
       const resourceMetadata = await discoverOAuthProtectedResourceMetadata(
         OPENCODE_SLACK_MCP_SERVER_URL,
