+++ packages/agent-core/tests/unit/opencode/task-manager.test.ts
@@ -28,9 +28,6 @@ describe('TaskManager', () => {
           platform: 'darwin' as NodeJS.Platform,
           isPackaged: false,
           tempPath: '/tmp',
-          getCliCommand: () => ({ command: 'opencode', args: [] }),
-          buildEnvironment: async (_taskId: string) => ({}),
-          buildCliArgs: async () => [],
         },
         defaultWorkingDirectory: '/home/user',
         isCliAvailable: async () => true,
